requirements.txt

torch
transformers
Pillow
streamlit
matplotlib
opencv-python
numpy
scipy
scikit-learn
